feat: Integrate upload-dif with debug-files upload command and preserve an alias
