+++ tests/integration/_cases/help/help.trycmd
@@ -28,7 +28,7 @@ OPTIONS:
     -V, --version                    Print version information
 
 SUBCOMMANDS:
-    debug-files        Locate, analyze or upload debug information files. [aliases: difutil]
+    debug-files        Locate, analyze or upload debug information files. [aliases: dif]
     help               Print this message or the help of the given subcommand(s)
     info               Print information about the Sentry server.
     issues             Manage issues in Sentry.
@@ -39,7 +39,6 @@ SUBCOMMANDS:
     repos              Manage repositories on Sentry.
     send-event         Send a manual event to Sentry.
     uninstall          Uninstall the sentry-cli executable.
-    upload-dif         Upload debugging information files.
     upload-proguard    Upload ProGuard mapping files to a project.
 
 ```
